working apollo loaders
