+++ js/actions/notifications.js
@@ -27,11 +27,17 @@
 const Platform = require('Platform');
 const VibrationIOS = require('VibrationIOS');
 const { updateInstallation } = require('./installation');
-const { loadNotifications } = require('./parse');
 const { loadSurveys } = require('./surveys');
 const { switchTab } = require('./navigation');
 
 import type { Action, ThunkAction } from './types';
+
+import gql from 'apollo-client/gql';
+import apollo from '../store/apollo';
+
+import loadApolloQuery from './apollo';
+
+
 
 type PushNotification = {
   foreground: boolean;
@@ -118,7 +124,28 @@ function markAllNotificationsAsSeen(): Action {
   };
 }
 
+function loadNotifications() : ThunkAction {
+  const query = apollo.query({
+    query: gql`
+      query viewer {
+        viewer {
+          notifications {
+            id
+            text
+            url
+            time
+          }
+        }
+      }
+    `,
+    forceFetch: false
+  });
+
+  return loadApolloQuery('LOADED_NOTIFICATIONS', query);
+}
+
 module.exports = {
+  loadNotifications,
   turnOnPushNotifications,
   storeDeviceToken,
   skipPushNotifications,
