+++ js/reducers/topics.js
@@ -30,8 +30,8 @@ type Action = { type: string; list: Array<any>; };
 function topics(state: State = [], action: Action): State {
   if (action.type === 'LOADED_SESSIONS') {
     var topicsMap = Object.create(null);
-    action.list.forEach((session) => {
-      var tags = session.get('tags') || [];
+    action.data.schedule.forEach((session) => {
+      var tags = session.tags || [];
       tags.forEach((tag) => {
         topicsMap[tag] = true;
       });
